Tidy up and fix tests for health
