+++ server/xpub-server/routes.js
@@ -1,13 +1,10 @@
 const logger = require('@pubsweet/logger')
-const config = require('config')
-const { db } = require('pubsweet-server')
-const AWS = require('aws-sdk')
+const { checkDataBase, checkS3 } = require('./health')
 
 const DATABASE_ERROR = 'Database Error'
 const S3_ERROR = 'S3 Error'
 const DEFAULT_ERROR = 'Internal Server Error'
-const SUCCESFUL_RESPONSE = 'pong'
-const S3_API_VERSION = '2006-03-01'
+const SUCCESSFUL_RESPONSE = 'pong'
 
 const nocache = (req, res, next) => {
   res.header('Cache-Control', 'private, no-cache, no-store, must-revalidate')
@@ -16,38 +13,13 @@ const nocache = (req, res, next) => {
   next()
 }
 
-const s3 = new AWS.S3({
-  ...config.aws.credentials,
-  ...config.aws.s3,
-  apiVersion: S3_API_VERSION,
-})
-
 let lastHealthStatus = null
-
-const checkS3 = () =>
-  new Promise(resolve => {
-    s3.listObjects({}, (err, response) => {
-      if (err) {
-        resolve(err)
-      } else {
-        resolve(response)
-      }
-    })
-  })
-
-const checkDataBase = () =>
-  new Promise(resolve => {
-    db.select('table_name')
-      .from('information_schema.tables')
-      .then(response => resolve(response))
-      .catch(error => resolve(error))
-  })
 
 module.exports = app => {
   app.get('/ping', nocache, async (req, res) => {
     const errors = []
     let statusCode = 200
-    let body = SUCCESFUL_RESPONSE
+    let body = SUCCESSFUL_RESPONSE
     let dbResponse
     let s3Response
     try {
